+++ NEWS.md
@@ -1,3 +1,5 @@
+# RandomWalker (development version)
+
 # RandomWalker 0.1.0
 
 ## Breaking Changes