+++ NEWS.md
@@ -15,6 +15,8 @@ maximum of a random walk.
 5. Fix #96 - Add Function `std_cum_mean_augment()` to calculate the cumulative
 mean of a random walk.
 6. Fix #113 - Add Function `get_attributes()` to get attributes without the `row.names`
+7. Fix #117 - Add `.interactive` parameter to `visualize_walks()` to allow for
+interactive plots.
 
 ## Minor Improvements and Fixes
 None