+++ NEWS.md
@@ -1,4 +1,4 @@
-# RandomWalker (development version)
+# RandomWalker 0.2.0
 
 ## Breaking Changes
 None